Flytt selve satsendring til egen transaksjon slik at vi kan endre på satsendringsobjektet i databasen selv om satsendringen går galt
